organize hook code
