+++ modules/firebase/src/hooks/useFirebaseApp.ts
@@ -1,0 +1,5 @@
+import { firebaseApp } from "../firebaseApp";
+
+export const useFirebaseApp = () => {
+  return firebaseApp;
+};
