+++ modules/firebase/src/hooks/useFirebaseUser.ts
